ref(crons): Show human readable cron tab text in sidebar
